requirements.txt
streamlit>=1.28.0
langchain-google-genai>=0.0.11
langgraph>=0.0.40
langsmith>=0.0.74
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=3.9.6
pydantic>=2.4.2
google-generativeai>=0.3.0
ddgs
geocoder
fpdf